cast pthread_self to long ints

+++ solutionOne.c
@@ -13,14 +13,14 @@ struct Semaphore mutexSem;
 //errors if name is only reader
 void *reader1(void *arg) {
     int thread_id = *((int *)arg);
-    printf("Reader %ld is waiting.\n", pthread_self());
+    printf("Reader %ld is waiting.\n", (unsigned long)pthread_self());
     new_wait(&mutexSem); //lock mutex and increase reader count
     readersCount++;
     if (readersCount == 1){
         new_wait(&resource_sem1); //if the first reader, lock resource as well
     }
     new_signal(&mutexSem); //unlock mutex for other readers to use
-    printf("Reader %ld is reading.\n", pthread_self());
+    printf("Reader %ld is reading.\n", (unsigned long)pthread_self());
 
     //reading the resource
 
@@ -30,7 +30,7 @@ void *reader1(void *arg) {
         new_signal(&resource_sem1);
     }
     new_signal(&mutexSem); //unlock mutex
-    printf("Reader %ld has finished reading.\n", pthread_self());
+    printf("Reader %ld has finished reading.\n", (unsigned long)pthread_self());
     return NULL;
 }
 
@@ -38,15 +38,15 @@ void *reader1(void *arg) {
 void *writer1(void *arg) {
     int thread_id = *((int *)arg);
     //wait to write
-    printf("Writer %ld is waiting.\n", pthread_self());
+    printf("Writer %ld is waiting.\n", (unsigned long)pthread_self());
     new_wait(&resource_sem1);
-    printf("Writer %ld is writing.\n", pthread_self());
+    printf("Writer %ld is writing.\n", (unsigned long)pthread_self());
 
     //simulate write
 
     //finish writing
     new_signal(&resource_sem1);
-    printf("Writer %ld has finished writing.\n", pthread_self());
+    printf("Writer %ld has finished writing.\n", (unsigned long)pthread_self());
     return NULL;
 }
